Fix mongoose env variable
env variables are always strings, therefore we have to parse RESETDBONLAUNCH to get a boolean

+++ loaders/mongoose.js
@@ -4,11 +4,11 @@
  * @returns {Promise<void>}
  */
 module.exports = async (mongoose) => {
-  console.log(`Connecting to mongo-db (${process.env.DBCONNECTION})...`)
-  await mongoose.connect(process.env.DBCONNECTION)
-  console.log('Connected to mongo-db.')
-  if (process.env.RESETDBONLAUNCH) {
-    await mongoose.connection.db.dropDatabase()
-    console.log('Reset db.')
-  }
-}
+	console.log(`Connecting to mongo-db (${process.env.DBCONNECTION})...`);
+	await mongoose.connect(process.env.DBCONNECTION);
+	console.log('Connected to mongo-db.');
+	if (process.env.RESETDBONLAUNCH && JSON.parse(process.env.RESETDBONLAUNCH)) {
+		await mongoose.connection.db.dropDatabase();
+		console.log('Reset db.');
+	}
+};
